backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-dateutil==2.9.0
cachetools==5.3.3
pytest==8.2.2
pytest-asyncio==0.23.7
